Backend/pyproject.toml
[project]
name = "explainability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=0.3.26",
    "langchain-aws>=0.2.31",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.71",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.8",
    "langchain-google-vertexai>=2.0.27",
    "langchain-groq>=0.3.6",
    "langchain-nvidia-ai-endpoints>=0.3.16",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-supervisor>=0.0.27",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.35.0",
]
